Safari-611.1.10/LayoutTests/js/arrowfunction-constructor-expected.txt

Tests for ES6 arrow function no constructor

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new simpleArrowFunction() threw exception TypeError: function is not a constructor (evaluating 'new simpleArrowFunction()').
PASS successfullyParsed is true

TEST COMPLETE
